Removed gcv/cf folder
This folder doesn't actually seem to be used any more (except for one utility function). Removing it clarifies the code structure.
